Add debug numpy backend + fixes
